Penguin CLean!

+++ OnEngine/src/OnEngine/ImGui/ImGuiLayer.cpp
@@ -76,7 +76,7 @@ namespace on
     {
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::GetInstance();
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
         float time = (float)glfwGetTime();
         io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
@@ -194,7 +194,7 @@ namespace on
     {
         ImGuiIO& io = ImGui::GetIO();
 
-        io.DisplaySize = ImVec2(event.GetWidth(), event.GetHeight());
+        io.DisplaySize = ImVec2((float)event.GetWidth(), (float)event.GetHeight());
 
         io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
